tap-to-mask/requirements.txt
numpy==1.19.5
opencv_python==4.5.3.56
Pillow==8.3.1
pycocotools==2.0.10
torch==1.10.2
transformers==4.57.3
